+++ CMakeLists.txt
@@ -8,7 +8,7 @@ set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 # 添加可执行文件，假设 main.cpp 位于项目根目录
-add_executable(My_tiny_STL main.cpp my_template.cpp my_operator.cpp)
+add_executable(My_tiny_STL main.cpp my_template.cpp my_operator.cpp my_reference.cpp)
 
 # 可选：为编译器启用更多警告
 if (CMAKE_CXX_COMPILER_ID MATCHES "GNU|Clang")
